Time: 12 ms (22.70%) | Memory: 45.5 MB (19.15%) - LeetSync

+++ 977-distinct-subsequences-ii/distinct-subsequences-ii.java
@@ -6,7 +6,7 @@ class Solution {
             return 1;
         }
         if(dp[n] != -2) return dp[n];
-        int  currans= (int)(2L*solve(n-1,prev)) % mod ;
+        int  currans= (2*solve(n-1,prev)) % mod ;
         int duplicates=0;
         if(prev[n-1]!=-1){
             duplicates= solve(prev[n-1], prev);
